chore: remove options namespace; move classes to Datatypes namespace
