+++ LotCoMPrinter/Models/Datatypes/RequiredFields.cs
@@ -1,6 +1,6 @@
 using CommunityToolkit.Mvvm.ComponentModel;
 
-namespace LotCoMPrinter.Models.Datasources;
+namespace LotCoMPrinter.Models.Datatypes;
 
 /// <summary>
 /// Provides structured control over the requirement of data fields in the variable set.
